+++ config/settings/base.py
@@ -3,6 +3,7 @@ Base settings to build other settings files upon.
 """
 
 import environ
+import topobank
 
 ROOT_DIR = environ.Path(__file__) - 3  # (topobank/config/settings/base.py - 3 = topobank/)
 APPS_DIR = ROOT_DIR.path('topobank')
@@ -76,6 +77,7 @@ THIRD_PARTY_APPS = [
     'bokeh',
     'termsandconditions',
     'storages',
+    'bootstrap_datepicker_plus', # for datepicker, see https://github.com/monim67/django-bootstrap-datepicker-plus
 ]
 LOCAL_APPS = [
     'topobank.users.apps.UsersAppConfig',
@@ -327,7 +329,7 @@ REST_FRAMEWORK = {
 }
 
 # Version number used in the GUI
-TOPOBANK_VERSION='0.1'
+TOPOBANK_VERSION=topobank.__version__
 
 #
 # Settings for authentication with ORCID