+++ topobank/manager/tests/utils.py
@@ -32,8 +32,14 @@ class TopographyFactory(factory.django.DjangoModelFactory):
 
     surface = factory.SubFactory(SurfaceFactory)
     name = factory.Sequence(lambda n: "topography-{}".format(n))
+    datafile = factory.django.FileField(from_path=str(settings.ROOT_DIR.path("topobank/manager/fixtures/10x10.txt")))
     data_source = 0
     measurement_date = factory.Sequence(lambda n: datetime.date(2019,1,1)+datetime.timedelta(days=n))
+    size_x = 512
+    size_y = None # set this if you need a 2D topography
+    size_editable = True
+    unit_editable = True
+    height_scale_editable = True
 
 #
 # Define fixtures